refactor: moving api.ts into lib folder
