+++ src/app/aspects/history/page.tsx
@@ -2,7 +2,7 @@
 
 import { AspectPoolHistory } from './aspect-pool-history'
 import { useEffect, useState } from 'react'
-import api from '@/utils/api'
+import api from '@/lib/api'
 import { Spinner } from '@/components/ui/spinner'
 
 export default function AspectPoolHistoryPage() {
