+++ src/app/lootrun/page.tsx
@@ -11,7 +11,7 @@ import Countdown from '@/components/custom/countdown'
 import { Spinner } from '@/components/ui/spinner'
 import Link from 'next/link'
 import { Button } from '@/components/ui/button'
-import api from '@/utils/api'
+import api from '@/lib/api'
 import { ArrowUpDown, CalendarSearch, CheckCircle2, Info, Search, Sparkles } from 'lucide-react'
 import { cn } from '@/lib/utils'
 import { Input } from '@/components/ui/input'
