+++ src/app/stats/page.tsx
@@ -5,7 +5,7 @@ import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
 import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "@/components/ui/table"
 import { Input } from "@/components/ui/input"
 // import { Pagination } from '@/components/ui/pagination'
-import api from '@/utils/api'
+import api from '@/lib/api'
 import { Button } from '@/components/ui/button'
 
 interface Guild {
