+++ src/app/stats/player/[playerName]/page.tsx
@@ -17,7 +17,7 @@ import { Dialog, DialogContent, DialogHeader, DialogTitle, DialogTrigger } from 
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select"
 import Image from 'next/image';
 import { MiscIcon } from '@/components/custom/WynnIcon';
-import api from '@/utils/api';
+import api from '@/lib/api';
 import { ChevronDown, Trophy } from 'lucide-react';
 import { Accordion, AccordionContent, AccordionItem, AccordionTrigger } from '@/components/ui/accordion';
 import { Separator } from '@/components/ui/separator';
